Add unmock-matcher as a package.
