+++ package.json
@@ -70,7 +70,8 @@
     "unmock-core": "file:packages/unmock-core",
     "unmock-jsdom": "file:packages/unmock-jsdom",
     "unmock-node": "file:packages/unmock-node",
-    "unmock-hash": "file:packages/unmock-hash"
+    "unmock-hash": "file:packages/unmock-hash",
+    "unmock-matcher": "file:packages/unmock-matcher"
   },
   "name": "unmock",
   "jest": {
