+++ packages/unmock-core/src/interfaces.ts
@@ -66,3 +66,20 @@ export interface IUnmockOptions {
   useInProduction?: boolean;
   mode?: Mode;
 }
+
+export interface ISerializedRequest {
+  body?: string;
+  method: string;
+  hostname: string;
+  pathname: string;
+}
+
+export interface ISerializedResponse {
+  body: any;
+  statusCode: number;
+}
+
+export interface IMock {
+  request: ISerializedRequest;
+  response: ISerializedResponse;
+}
