+++ packages/unmock-matcher/package.json
@@ -20,11 +20,9 @@
     "@types/uuid": "^3.4.4",
     "content-type": "^1.0.4",
     "debug": "^4.1.1",
-    "io-ts": "^1.8.6",
-    "io-ts-reporters": "^0.0.21",
     "lodash": "^4.17.11",
-    "object-hash": "^1.3.1",
     "querystring": "^0.2.0",
-    "readable-stream": "^3.4.0"
+    "readable-stream": "^3.4.0",
+    "unmock-core": "file:../unmock-core"
   }
 }
