+++ tsconfig.build.json
@@ -10,6 +10,7 @@
     { "path": "./packages/unmock-core" },
     { "path": "./packages/unmock-hash" },
     { "path": "./packages/unmock-jsdom" },
-    { "path": "./packages/unmock-node" }
+    { "path": "./packages/unmock-node" },
+    { "path": "./packages/unmock-matcher" }
   ]
 }
