+++ tsconfig.test.json
@@ -10,6 +10,7 @@
     { "path": "./packages/unmock-core/src/__tests__" },
     { "path": "./packages/unmock-hash/src/__tests__" },
     { "path": "./packages/unmock-jsdom/src/__tests__" },
-    { "path": "./packages/unmock-node/src/__tests__" }
+    { "path": "./packages/unmock-node/src/__tests__" },
+    { "path": "./packages/unmock-matcher/src/__tests__" }
   ]
 }
